requirements.txt
PySide6>=6.10.0
Pillow>=10.0.0

